Grasping at straws, maybe this will work?

+++ src/lib/lofar_udp_backends.hpp
@@ -387,7 +387,7 @@ void inline udp_stokes(long iLoop, char *inputPortData, O **outputData,  long la
 
 template <typename I, typename O, StokesFuncType stokesFunc, int factor>
 void inline udp_stokesDecimation(long iLoop, char *inputPortData, O **outputData, long lastInputPacketOffset, long packetOutputLength, int timeStepSize, int totalBeamlets, int upperBeamlet, int cumulativeBeamlets, int baseBeamlet) {
-	long outputPacketOffset = (long) ((float) iLoop * packetOutputLength / sizeof(O));
+	long outputPacketOffset = iLoop * packetOutputLength / sizeof(O);
 	long tsInOffset, tsOutOffset;
 	O tempVal;
 
@@ -454,7 +454,7 @@ void inline udp_fullStokes(long iLoop, char *inputPortData, O **outputData,  lon
 
 template <typename I, typename O, int factor>
 void inline udp_fullStokesDecimation(long iLoop, char *inputPortData, O **outputData, long lastInputPacketOffset, long packetOutputLength, int timeStepSize, int totalBeamlets, int upperBeamlet, int cumulativeBeamlets, int baseBeamlet) {
-	long outputPacketOffset = (long) ((float) iLoop * packetOutputLength / sizeof(O));
+	long outputPacketOffset = iLoop * packetOutputLength / sizeof(O);
 	long tsInOffset, tsOutOffset;
 	O tempValI, tempValQ, tempValU, tempValV;
 
@@ -533,7 +533,7 @@ void inline udp_usefulStokes(long iLoop, char *inputPortData, O **outputData,  l
 
 template <typename I, typename O, int factor>
 void inline udp_usefulStokesDecimation(long iLoop, char *inputPortData, O **outputData, long lastInputPacketOffset, long packetOutputLength, int timeStepSize, int totalBeamlets, int upperBeamlet, int cumulativeBeamlets, int baseBeamlet) {
-	long outputPacketOffset = (long) ((float) iLoop * packetOutputLength / sizeof(O));
+	long outputPacketOffset = iLoop * packetOutputLength / sizeof(O);
 	long tsInOffset, tsOutOffset;
 	O tempValI, tempValV;
